RAGapplication/pyproject.toml
[project]
name = "ragapplication"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.49",
    "langchain-groq>=0.2.1",
    "langchain-openai>=0.2.12",
    "langchain-pinecone>=0.2.0",
    "langchain>=0.3.22",
    "langsmith>=0.2.2",
    "pandas>=2.2.3",
    "pdfreader>=0.1.15",
    "pinecone>=5.3.1",
    "pypdf>=5.1.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "weave>=0.51.23",
    "cohere>=5.13.3",
    "datetime>=5.5",
    "wikipedia>=1.4.0",
    "llama-index-agent-openai==0.2.9",
    "llama-index-cli==0.1.13",
    "llama-index-llms-openai==0.1.31",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai==2.0.11",
    "pyowm>=3.3.0",
    "docling>=2.12.0",
    "docling-core>=2.10.0",
    "rank-bm25>=0.2.2",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "langchain-experimental>=0.3.4",
    "langgraph>=0.2.60",
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "weaviate-client>=4.10.2",
    "langchain-weaviate>=0.0.3",
    "scikit-learn>=1.6.0",
    "duckduckgo-search>=7.2.0",
    "langchain-ollama>=0.2.2",
    "langchain-ai21>=1.0.1",
    "sseclient-py>=1.8.0",
    "langchain-mongodb>=0.4.0",
    "langchain-chroma>=0.1.2",
    "lark>=1.2.2",
    "opik>=1.4.5",
    "ragas>=0.2.12",
    "autoflake>=2.3.1",
    "flake8>=7.1.1",
    "black>=25.1.0",
    "mypy>=1.14.1",
    "isort>=6.0.0",
    "tavily-python>=0.5.1",
    "email-validator>=2.2.0",
    "langchain-linkup>=0.1.5",
    "pdfplumber>=0.11.5",
    "deep-translator>=1.11.4",
    "wayback>=0.4.5",
    "waybackpy>=3.0.6",
    "requests-html>=0.10.0",
    "exa-py>=1.9.0",
    "langchain-exa>=0.2.1",
    "boto3>=1.37.17",
    "google-news-feed>=1.1.0",
    "googlenewsdecoder>=0.1.7",
    "newsapi-python>=0.2.7",
    "gnews>=0.4.1",
    "serpapi>=0.1.5",
    "google-search-results>=2.4.2",
    "schedule>=1.2.2",
    "apscheduler>=3.11.0",
    "fastapi-utilities>=0.3.1",
    "colorlog>=6.9.0",
    "loguru>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-env>=1.1.5,<2.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 88
skip-string-normalization = true
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
